+++ app/core/src/main/java/com/topjohnwu/magisk/core/model/module/Module.kt
@@ -7,7 +7,7 @@ abstract class Module : Comparable<Module> {
         protected set
     abstract var version: String
         protected set
-    abstract var versionCode: Int
+    abstract var versionCode: Long
         protected set
 
     override operator fun compareTo(other: Module) = id.compareTo(other.id)
